+++ apps/elf-eval/tests/real_world_job_benchmark/manifest_summary.rs
@@ -1,3 +1,6 @@
+mod manifest_summary_scenario;
+mod manifest_summary_status;
+
 use serde_json::Value;
 
 pub(super) fn assert_external_adapter_manifest_summary(report: &Value) {
@@ -48,198 +51,6 @@ pub(super) fn assert_external_adapter_manifest_summary(report: &Value) {
 		Some(14)
 	);
 
-	assert_external_adapter_manifest_status_summary(report);
-	assert_external_adapter_manifest_scenario_summary(report);
-}
-
-fn assert_external_adapter_manifest_status_summary(report: &Value) {
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/overall_status_counts/pass")
-			.and_then(Value::as_u64),
-		Some(4)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/overall_status_counts/wrong_result")
-			.and_then(Value::as_u64),
-		Some(6)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/overall_status_counts/lifecycle_fail")
-			.and_then(Value::as_u64),
-		Some(1)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/overall_status_counts/incomplete")
-			.and_then(Value::as_u64),
-		Some(0)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/overall_status_counts/blocked")
-			.and_then(Value::as_u64),
-		Some(10)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/overall_status_counts/not_encoded")
-			.and_then(Value::as_u64),
-		Some(5)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/capability_status_counts/mocked")
-			.and_then(Value::as_u64),
-		Some(1)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/capability_status_counts/unsupported")
-			.and_then(Value::as_u64),
-		Some(6)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/suite_status_counts/blocked")
-			.and_then(Value::as_u64),
-		Some(29)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/suite_status_counts/pass")
-			.and_then(Value::as_u64),
-		Some(27)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/suite_status_counts/incomplete")
-			.and_then(Value::as_u64),
-		Some(0)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/suite_status_counts/not_encoded")
-			.and_then(Value::as_u64),
-		Some(37)
-	);
-}
-
-fn assert_external_adapter_manifest_scenario_summary(report: &Value) {
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_status_counts/real")
-			.and_then(Value::as_u64),
-		Some(0)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_status_counts/mocked")
-			.and_then(Value::as_u64),
-		Some(0)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_status_counts/unsupported")
-			.and_then(Value::as_u64),
-		Some(3)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_status_counts/blocked")
-			.and_then(Value::as_u64),
-		Some(24)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_status_counts/incomplete")
-			.and_then(Value::as_u64),
-		Some(5)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_status_counts/wrong_result")
-			.and_then(Value::as_u64),
-		Some(6)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_status_counts/lifecycle_fail")
-			.and_then(Value::as_u64),
-		Some(1)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_status_counts/pass")
-			.and_then(Value::as_u64),
-		Some(23)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_status_counts/not_encoded")
-			.and_then(Value::as_u64),
-		Some(13)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_position_counts/wins")
-			.and_then(Value::as_u64),
-		Some(10)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_position_counts/ties")
-			.and_then(Value::as_u64),
-		Some(11)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_position_counts/loses")
-			.and_then(Value::as_u64),
-		Some(1)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_position_counts/untested")
-			.and_then(Value::as_u64),
-		Some(53)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_outcome_counts/win")
-			.and_then(Value::as_u64),
-		Some(10)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_outcome_counts/tie")
-			.and_then(Value::as_u64),
-		Some(11)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_outcome_counts/loss")
-			.and_then(Value::as_u64),
-		Some(1)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_outcome_counts/not_tested")
-			.and_then(Value::as_u64),
-		Some(19)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_outcome_counts/blocked")
-			.and_then(Value::as_u64),
-		Some(29)
-	);
-	assert_eq!(
-		report
-			.pointer("/external_adapters/summary/scenario_outcome_counts/non_goal")
-			.and_then(Value::as_u64),
-		Some(5)
-	);
+	manifest_summary_status::assert_external_adapter_manifest_status_summary(report);
+	manifest_summary_scenario::assert_external_adapter_manifest_scenario_summary(report);
 }
