+++ apps/elf-eval/tests/real_world_job_benchmark/memory_evolution.rs
@@ -1,215 +1,2 @@
-use std::{env, fs, process};
-
-use color_eyre::Result;
-use serde_json::Value;
-
-use crate::support;
-
-#[test]
-fn memory_evolution_fixtures_report_temporal_and_staleness_metrics() -> Result<()> {
-	let report = support::run_json_report_from(support::evolution_fixture_dir())?;
-
-	assert_eq!(report.pointer("/summary/job_count").and_then(Value::as_u64), Some(5));
-	assert_eq!(report.pointer("/summary/encoded_suite_count").and_then(Value::as_u64), Some(1));
-	assert_eq!(report.pointer("/summary/pass").and_then(Value::as_u64), Some(5));
-	assert_eq!(report.pointer("/summary/not_encoded").and_then(Value::as_u64), Some(0));
-	assert_eq!(report.pointer("/summary/stale_answer_count").and_then(Value::as_u64), Some(0));
-	assert_eq!(
-		report.pointer("/summary/conflict_detection_count").and_then(Value::as_u64),
-		Some(5)
-	);
-	assert_eq!(
-		report.pointer("/summary/update_rationale_available_count").and_then(Value::as_u64),
-		Some(5)
-	);
-	assert_eq!(
-		report.pointer("/summary/temporal_validity_not_encoded_count").and_then(Value::as_u64),
-		Some(0)
-	);
-	assert_eq!(
-		report.pointer("/summary/history_readback_encoded_count").and_then(Value::as_u64),
-		Some(1)
-	);
-	assert_eq!(
-		report.pointer("/evolution/temporal_validity_not_encoded_count").and_then(Value::as_u64),
-		Some(0)
-	);
-	assert_eq!(
-		report.pointer("/evolution/history_readback_encoded_count").and_then(Value::as_u64),
-		Some(1)
-	);
-
-	let suites = support::array_at(&report, "/suites")?;
-	let memory_evolution = support::find_by_field(suites, "/suite_id", "memory_evolution")?;
-
-	assert_eq!(memory_evolution.pointer("/status").and_then(Value::as_str), Some("pass"));
-	assert_eq!(
-		memory_evolution.pointer("/temporal_validity_not_encoded_count").and_then(Value::as_u64),
-		Some(0)
-	);
-	assert_eq!(
-		memory_evolution.pointer("/history_readback_encoded_count").and_then(Value::as_u64),
-		Some(1)
-	);
-
-	let jobs = support::array_at(&report, "/jobs")?;
-	let preference_job =
-		support::find_by_field(jobs, "/job_id", "memory-evolution-preference-001")?;
-	let relation_job =
-		support::find_by_field(jobs, "/job_id", "memory-evolution-relation-temporal-001")?;
-
-	assert_eq!(
-		preference_job.pointer("/evolution/history_readback_encoded").and_then(Value::as_bool),
-		Some(true)
-	);
-	assert!(support::array_contains_str(preference_job, "/evolution/history_event_types", "add")?);
-	assert!(support::array_contains_str(
-		preference_job,
-		"/evolution/history_event_types",
-		"update"
-	)?);
-	assert!(support::array_contains_str(
-		preference_job,
-		"/evolution/history_event_types",
-		"ignore"
-	)?);
-	assert_eq!(
-		preference_job
-			.pointer("/evolution/history_requires_note_version_links")
-			.and_then(Value::as_bool),
-		Some(true)
-	);
-	assert_eq!(
-		preference_job.pointer("/evolution/selected_current_evidence/0").and_then(Value::as_str),
-		Some("pref-current-concise-rationale")
-	);
-	assert_eq!(
-		preference_job.pointer("/evolution/selected_historical_evidence/0").and_then(Value::as_str),
-		Some("pref-old-terse-bullets")
-	);
-	assert_eq!(
-		preference_job.pointer("/evolution/selected_rationale_evidence/0").and_then(Value::as_str),
-		Some("pref-update-rationale")
-	);
-	assert_eq!(relation_job.pointer("/status").and_then(Value::as_str), Some("pass"));
-	assert_eq!(
-		relation_job.pointer("/evolution/temporal_validity_not_encoded").and_then(Value::as_bool),
-		Some(false)
-	);
-	assert_eq!(
-		relation_job.pointer("/evolution/temporal_validity_encoded").and_then(Value::as_bool),
-		Some(true)
-	);
-
-	let follow_ups = support::array_at(&report, "/follow_ups")?;
-
-	assert!(follow_ups.is_empty());
-
-	Ok(())
-}
-
-#[test]
-fn memory_evolution_conflict_still_fails_when_selected_evidence_is_not_narrated() -> Result<()> {
-	let fixture_path =
-		support::evolution_fixture_dir().join("preference_changed_current_vs_historical.json");
-	let mut fixture = serde_json::from_str::<Value>(&fs::read_to_string(fixture_path)?)?;
-
-	support::set_json_pointer(
-		&mut fixture,
-		"/corpus/adapter_response/answer/evidence_ids",
-		serde_json::json!([
-			"pref-current-concise-rationale",
-			"pref-old-terse-bullets",
-			"pref-update-rationale"
-		]),
-	)?;
-	support::set_json_pointer(
-		&mut fixture,
-		"/corpus/adapter_response/answer/claims",
-		serde_json::json!([
-			{
-				"claim_id": "current_preference",
-				"text": "Use concise prose with explicit evidence before bullets.",
-				"evidence_ids": ["pref-current-concise-rationale", "pref-update-rationale"],
-				"confidence": "high"
-			},
-			{
-				"claim_id": "preference_update_rationale",
-				"text": "The preference changed because terse bullets hid rationale.",
-				"evidence_ids": ["pref-update-rationale"],
-				"confidence": "high"
-			}
-		]),
-	)?;
-
-	let temp_dir =
-		env::temp_dir().join(format!("elf-real-world-memory-conflict-test-{}", process::id()));
-
-	fs::create_dir_all(&temp_dir)?;
-	fs::write(temp_dir.join("conflict.json"), serde_json::to_vec_pretty(&fixture)?)?;
-
-	let report = support::run_json_report_from(temp_dir)?;
-	let jobs = support::array_at(&report, "/jobs")?;
-	let job = support::find_by_field(jobs, "/job_id", "memory-evolution-preference-001")?;
-
-	assert_eq!(job.pointer("/status").and_then(Value::as_str), Some("wrong_result"));
-	assert_eq!(job.pointer("/evolution/conflict_detection_count").and_then(Value::as_u64), Some(0));
-	assert!(support::array_contains_str(
-		job,
-		"/evolution/selected_but_not_narrated_evidence",
-		"pref-old-terse-bullets"
-	)?);
-
-	Ok(())
-}
-
-#[test]
-fn memory_evolution_counts_stale_answer_when_old_fact_is_answered_as_current() -> Result<()> {
-	let fixture_path =
-		support::evolution_fixture_dir().join("preference_changed_current_vs_historical.json");
-	let mut fixture = serde_json::from_str::<Value>(&fs::read_to_string(fixture_path)?)?;
-
-	support::set_json_pointer(
-		&mut fixture,
-		"/corpus/adapter_response/answer/content",
-		Value::String(
-			"Use terse bullet-only benchmark updates as the current preference.".to_string(),
-		),
-	)?;
-	support::set_json_pointer(
-		&mut fixture,
-		"/corpus/adapter_response/answer/evidence_ids",
-		serde_json::json!(["pref-old-terse-bullets"]),
-	)?;
-	support::set_json_pointer(
-		&mut fixture,
-		"/corpus/adapter_response/answer/claims",
-		serde_json::json!([
-			{
-				"claim_id": "current_preference",
-				"text": "Use terse bullet-only benchmark updates as the current preference.",
-				"evidence_ids": ["pref-old-terse-bullets"],
-				"confidence": "high"
-			}
-		]),
-	)?;
-
-	let temp_dir =
-		env::temp_dir().join(format!("elf-real-world-memory-stale-test-{}", process::id()));
-
-	fs::create_dir_all(&temp_dir)?;
-	fs::write(temp_dir.join("stale_preference.json"), serde_json::to_vec_pretty(&fixture)?)?;
-
-	let report = support::run_json_report_from(temp_dir)?;
-
-	assert_eq!(report.pointer("/summary/stale_answer_count").and_then(Value::as_u64), Some(1));
-	assert_eq!(report.pointer("/summary/wrong_result").and_then(Value::as_u64), Some(1));
-
-	let jobs = support::array_at(&report, "/jobs")?;
-	let job = support::find_by_field(jobs, "/job_id", "memory-evolution-preference-001")?;
-
-	assert_eq!(job.pointer("/status").and_then(Value::as_str), Some("wrong_result"));
-	assert_eq!(job.pointer("/evolution/stale_answer_count").and_then(Value::as_u64), Some(1));
-
-	Ok(())
-}
+mod memory_evolution_mutations;
+mod memory_evolution_scoring;
