+++ apps/elf-eval/tests/real_world_job_benchmark/memory_evolution_mutations.rs
@@ -1,0 +1,111 @@
+use std::{env, fs, process};
+
+use color_eyre::Result;
+use serde_json::Value;
+
+use crate::support;
+
+#[test]
+fn memory_evolution_conflict_still_fails_when_selected_evidence_is_not_narrated() -> Result<()> {
+	let fixture_path =
+		support::evolution_fixture_dir().join("preference_changed_current_vs_historical.json");
+	let mut fixture = serde_json::from_str::<Value>(&fs::read_to_string(fixture_path)?)?;
+
+	support::set_json_pointer(
+		&mut fixture,
+		"/corpus/adapter_response/answer/evidence_ids",
+		serde_json::json!([
+			"pref-current-concise-rationale",
+			"pref-old-terse-bullets",
+			"pref-update-rationale"
+		]),
+	)?;
+	support::set_json_pointer(
+		&mut fixture,
+		"/corpus/adapter_response/answer/claims",
+		serde_json::json!([
+			{
+				"claim_id": "current_preference",
+				"text": "Use concise prose with explicit evidence before bullets.",
+				"evidence_ids": ["pref-current-concise-rationale", "pref-update-rationale"],
+				"confidence": "high"
+			},
+			{
+				"claim_id": "preference_update_rationale",
+				"text": "The preference changed because terse bullets hid rationale.",
+				"evidence_ids": ["pref-update-rationale"],
+				"confidence": "high"
+			}
+		]),
+	)?;
+
+	let temp_dir =
+		env::temp_dir().join(format!("elf-real-world-memory-conflict-test-{}", process::id()));
+
+	fs::create_dir_all(&temp_dir)?;
+	fs::write(temp_dir.join("conflict.json"), serde_json::to_vec_pretty(&fixture)?)?;
+
+	let report = support::run_json_report_from(temp_dir)?;
+	let jobs = support::array_at(&report, "/jobs")?;
+	let job = support::find_by_field(jobs, "/job_id", "memory-evolution-preference-001")?;
+
+	assert_eq!(job.pointer("/status").and_then(Value::as_str), Some("wrong_result"));
+	assert_eq!(job.pointer("/evolution/conflict_detection_count").and_then(Value::as_u64), Some(0));
+	assert!(support::array_contains_str(
+		job,
+		"/evolution/selected_but_not_narrated_evidence",
+		"pref-old-terse-bullets"
+	)?);
+
+	Ok(())
+}
+#[test]
+fn memory_evolution_counts_stale_answer_when_old_fact_is_answered_as_current() -> Result<()> {
+	let fixture_path =
+		support::evolution_fixture_dir().join("preference_changed_current_vs_historical.json");
+	let mut fixture = serde_json::from_str::<Value>(&fs::read_to_string(fixture_path)?)?;
+
+	support::set_json_pointer(
+		&mut fixture,
+		"/corpus/adapter_response/answer/content",
+		Value::String(
+			"Use terse bullet-only benchmark updates as the current preference.".to_string(),
+		),
+	)?;
+	support::set_json_pointer(
+		&mut fixture,
+		"/corpus/adapter_response/answer/evidence_ids",
+		serde_json::json!(["pref-old-terse-bullets"]),
+	)?;
+	support::set_json_pointer(
+		&mut fixture,
+		"/corpus/adapter_response/answer/claims",
+		serde_json::json!([
+			{
+				"claim_id": "current_preference",
+				"text": "Use terse bullet-only benchmark updates as the current preference.",
+				"evidence_ids": ["pref-old-terse-bullets"],
+				"confidence": "high"
+			}
+		]),
+	)?;
+
+	let temp_dir =
+		env::temp_dir().join(format!("elf-real-world-memory-stale-test-{}", process::id()));
+
+	fs::create_dir_all(&temp_dir)?;
+	fs::write(temp_dir.join("stale_preference.json"), serde_json::to_vec_pretty(&fixture)?)?;
+
+	let report = support::run_json_report_from(temp_dir)?;
+
+	assert_eq!(report.pointer("/summary/stale_answer_count").and_then(Value::as_u64), Some(1));
+	assert_eq!(report.pointer("/summary/wrong_result").and_then(Value::as_u64), Some(1));
+
+	let jobs = support::array_at(&report, "/jobs")?;
+	let job = support::find_by_field(jobs, "/job_id", "memory-evolution-preference-001")?;
+
+	assert_eq!(job.pointer("/status").and_then(Value::as_str), Some("wrong_result"));
+	assert_eq!(job.pointer("/evolution/stale_answer_count").and_then(Value::as_u64), Some(1));
+
+	Ok(())
+}
